pyproject.toml
[project]
name = "ai-sessions"
version = "0.1.0"
description = "Thin Python CLI + FastAPI HTTP API to call, manage, and view sessions across claude, codex, and opencode."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "typer>=0.13",
    "python-dotenv>=1.0",
    "claude-agent-sdk>=0.0.10",
    # OpenAI Codex Python SDK. PyPI: codex-app-server-sdk; import: codex_app_server.
    # Requires Python >=3.12.
    "codex-app-server-sdk>=0.3.0",
]

[project.scripts]
ai-sessions = "ai_sessions.cli:app"
ais = "ai_sessions.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_sessions"]
